chore: tailwind local styles
