+++ .storybook/preview.tsx
@@ -1,8 +1,6 @@
 import React from "react";
 import { kebabCase } from "lodash";
 
-import "./global.css";
-
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
   controls: {
